backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
pydantic-settings==2.10.1
email-validator==2.3.0
PyJWT==2.10.1
openai==1.107.1
